backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djongo==1.3.6
pymongo==4.6.1
python-dotenv==1.0.1
django-cors-headers==4.3.1
channels==4.0.0
daphne==4.0.0
Pillow==10.2.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
python-magic==0.4.27
django-storages==1.14.2
boto3==1.34.34
django-cleanup==8.0.0 